% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_sys}
\alias{app_sys}
\title{Access files in the current app}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Character vectors specifying the path within the \code{inst/app/www} folder.}
}
\value{
The full path to the file.
}
\description{
Access files in the current app
}
\examples{
\dontrun{
  app_sys("js", "app.js")
}
}
