% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_codes.R
\name{get_species_codes}
\alias{get_species_codes}
\title{Clean and Process Species Codes}
\usage{
get_species_codes(
  species_path = external_files()$species_codes$path,
  metadata_path = external_files()$species_metadata$path,
  drop_subspecies = TRUE
)
}
\arguments{
\item{species_path}{Path to the DBF file containing species codes.}

\item{metadata_path}{Path to the CSV file containing metadata.}

\item{drop_subspecies}{Logical, whether to drop subspecies from the resulting data frame (default: TRUE).}
}
\value{
A cleaned data frame of species codes with optional subspecies filtered out.
}
\description{
Functions to clean and process species codes from a DBF file, join them with metadata, and optionally drop subspecies.
}
\details{
Clean and Process Species Codes

This file contains functions to clean and process species codes from a DBF file,
join them with metadata, and optionally drop subspecies from the output.
}
\examples{
\dontrun{
species_ref <- get_species_codes(
 species_path = "data/CodesEspeces.dbf", 
 metadata_path = "data/metadata_species.csv", 
 drop_subspecies = TRUE
)
}
}
