% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_codes.R
\name{get_species_rank}
\alias{get_species_rank}
\title{Get species rank}
\usage{
get_species_rank(scientific_name)
}
\arguments{
\item{scientific_name}{The scientific name of the species to be ranked}
}
\value{
A character string indicating the rank: "genre", "subspecies", or "species".
}
\description{
Get species rank based on scientific name
}
\examples{
get_species_rank("Aquila chrysaetos")

}
