% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_datasets.R
\name{load_all_datasets}
\alias{load_all_datasets}
\title{Load and process all datasets}
\usage{
load_all_datasets(combine = TRUE)
}
\arguments{
\item{combine}{Logical, if TRUE (default), combines all datasets into a single \code{data.frame}.
If FALSE, returns a named list of individual \code{data.frames}.}
}
\value{
A combined \code{data.frame} or a list of \code{data.frames} containing all processed datasets.
}
\description{
This function sequentially calls all the \verb{load_*} dataset processing functions
and combines their outputs into a single \code{data.frame} or a list of \code{data.frames}.
}
\examples{
\dontrun{
all_data <- load_all_datasets()
}
}
