% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_inventaire_aerien_Nunavik.R
\name{load_inventaire_aerien_nunavik}
\alias{load_inventaire_aerien_nunavik}
\title{Load and process inventaire_aerien_nunavik.csv dataset}
\usage{
load_inventaire_aerien_nunavik()
}
\value{
A processed \code{data.frame} with standardized columns and integrated species codes.
}
\description{
This function loads and processes the "inventaire_aerien_nunavik.gdb" dataset from a predefined external source.
It validates the file and columns, applies transformations, and integrates species codes using a reference table.
}
\examples{
\dontrun{
processed_macreuse <- load_macreuse()
}
}
