% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_datasets_folder}
\alias{set_datasets_folder}
\title{Set datasets folder path}
\usage{
set_datasets_folder(path)
}
\arguments{
\item{path}{Path to the datasets folder (must end with "/")}
}
\description{
Set datasets folder path
}
