% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgenceAviR-package.R
\docType{package}
\name{urgenceAviR-package}
\alias{urgenceAviR}
\alias{urgenceAviR-package}
\title{urgenceAviR: urgenceAviR}
\description{
Un package conçu pour rassembler et intégrer les relevés d'oiseaux. La nouvelle base de données créée est utilisée pour évaluer le niveau de menace environnementale lors d'incident.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://insileco.io/urgenceAviR/}
}

}
\author{
\strong{Maintainer}: inSileco \email{info@insileco.io}

Authors:
\itemize{
  \item Matthieu Beaumont \email{matthieu.beaumont@ec.gc.ca}
}

Other contributors:
\itemize{
  \item Clara Casabona [contributor]
}

}
\keyword{internal}
