% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebird.R
\name{get_ebird}
\alias{get_ebird}
\title{Retrieve eBird Data}
\usage{
get_ebird(
  path = NULL,
  species = NULL,
  year = NULL,
  month = NULL,
  extent = NULL,
  ...
)
}
\arguments{
\item{path}{Character string specifying the file path to the eBird GeoDatabase file (.gdb).
If \code{NULL}, uses the default path from \code{external_files()$ebird_data$path}.}

\item{species}{Character vector of species names to filter (e.g., "Snow Goose").
If \code{NULL}, the query will not filter by species.}

\item{year}{Integer vector of years to filter (e.g., \code{2012:2014}).
If \code{NULL}, the query will not filter by year.}

\item{month}{Integer vector of months to filter (e.g., \code{4:5}). If \code{NULL}, the query
will filter by year only.}

\item{extent}{An \code{sf} or \code{terra} spatial object defining the spatial extent to filter data.
If \code{NULL}, the query will not apply spatial filtering.}

\item{...}{Additional arguments passed to the SQL query construction function
(\code{compose_query}). This allows for further customization of query conditions.}
}
\value{
A \code{SpatVector} object containing the filtered eBird data.
}
\description{
Filters and retrieves data from an eBird GeoDatabase file (.gdb) using a SQL query.
}
\examples{
\dontrun{
# Example: Retrieve data for Snow Goose in April and May of 2012-2014
get_ebird(
  path = "path/to/ebird.gdb",
  species = "Snow Goose",
  year = 2012:2014,
  month = 4:5
)

# Using default path from external_files
get_ebird(species = "Snow Goose", year = 2020)
}

}
